+++ app/src/main/java/com/app/music_app/exercises/logic/CompareExerciseLogic.kt
@@ -1,32 +1,20 @@
-package com.app.music_app.tasks.logic
+package com.app.music_app.exercises.logic
 
 import android.content.Context
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.app.music_app.music_player.instruments.VirtualPiano
-import com.app.music_app.tasks.pages.ChooseTaskPage
-import com.app.music_app.tasks.pages.ResultsPage
-import com.app.music_app.view.colors.AppColor
+import com.app.music_app.exercises.pages.ChooseTaskPage
+import com.app.music_app.exercises.pages.ResultsPage
 import com.app.music_app.view.piano_keyboard.PianoKeyboard
 import com.app.music_app.view.progress_bar.TaskProgressBar
 import com.musiclib.notes.range.NoteRange
