+++ app/src/main/java/com/app/music_app/exercises/pages/IntervalPlayPage.kt
@@ -1,9 +1,6 @@
-package com.app.music_app.tasks.pages
+package com.app.music_app.exercises.pages
 
-import android.Manifest
-import android.app.Activity
 import android.content.Context
-import android.content.pm.PackageManager
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -14,22 +11,15 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import com.app.music_app.music_player.MelodyPlayer
-import com.app.music_app.music_player.interfaces.AbstractInstrument
 import com.app.music_app.pitch_analyzer.NoteDetector
 import com.app.music_app.view.colors.AppColor
 import com.app.music_app.view.paino_box.PianoBox
